Made AbstractUndoRedoAction
The same code was written in both UndoAction and RedoAction, so an abstract class with the generic code was created and both extend this abstract class.
